Rename Evaluator.
